code optimization

+++ resolvers/coupon.js
@@ -1,24 +1,32 @@
 const Coupon = require("../models/Coupon");
 const {
   isEmpty,
-  putError,
   checkError,
-  imageUpload,
-  imageUnlink,
   checkToken
 } = require("../config/helpers");
 const validate = require("../validations/coupon");
-const errorRES = require("../error");
+const Messages = require("../config/messages");
 
 module.exports = {
   Query: {
     coupons: async (root, args) => {
       try {
         const coupons = await Coupon.find({});
-        return coupons || [];
+        return {
+          message: {
+            message: Messages.RESULT_FOUND.replace(":item", "Coupon"),
+            success: true,
+          },
+          data: coupons,
+        };
       } catch (error) {
-        throw new Error("Something went wrong.");
-      }a
+        return {
+          message: {
+            message: Messages.RETRIEVE_ERROR.replace(":item", "Coupon"),
+            success: false,
+          },
+        };
+      }
     },
     // get all coupons with pagination................
 
@@ -58,43 +66,84 @@ module.exports = {
         return {
           meta_data: { totalCount: total, page: pageNumber },
           data: edges,
-          message: { message: `${errorRES.RETRIEVE_ERROR} coupon`, status: 200 },
+          message: {
+            message: Messages.RETRIEVE_ERROR.replace(":item", "Coupon"),
+            success: false,
+          },
         };
       } else {
         return {
           pagination: { totalCount: total, page: pageNumber },
           data: edges,
-          message: { message: "Coupons list fetched", status: 200 },
+          message: {
+            message: Messages.RESULT_FOUND.replace(":item", "Coupon"),
+            success: true,
+          },
         };
       }
     },
     coupon: async (root, args) => {
+      if (!args.id) {
+        return {
+          message: {
+            message: Messages.ID_ERROR.replace(":item", "Coupon"),
+            success: false,
+          },
+        };
+      }
       try {
         const coupon = await Coupon.findById(args.id);
         if (!coupon) {
-          throw putError("Coupon not found");
+          return {
+            message: {
+              message: Messages.NOT_EXIST.replace(":item", "Coupon"),
+              success: false,
+            },
+          };
         }
-        return coupon;
+        return {
+          message: {
+            message: Messages.RESULT_FOUND.replace(":item", "Coupon"),
+            success: true,
+          },
+          data: coupon,
+        };
       } catch (error) {
         error = checkError(error);
-        throw new Error(error.custom_message);
+        return {
+          message: {
+            message: Messages.RETRIEVE_ERROR.replace(":item", "Coupon"),
+            success: false,
+          },
+        };
       }
     }
   },
   Mutation: {
     addCoupon: async (root, args, { id }) => {
+      if (!id) {
+        return {
+          message: Messages.TOKEN_REQ.replace(":item", "Coupon"),
+          success: false,
+        };
+      }
       checkToken(id);
       try {
-        // Check Validation
         const errors = validate("addCoupon", args);
         if (!isEmpty(errors)) {
-          throw putError(errors);
+          return {
+            message: errors,
+            success: false,
+          };
         }
 
         const coupon = await Coupon.findOne({ code: args.code });
 
         if (coupon) {
-          throw putError("Coupon code already exist.");
+          return {
+            message: Messages.DUPLICATE.replace(":item", "CouponCode"),
+            success: false,
+          };
         } else {
           const newCoupon = new Coupon({
             code: args.code,
@@ -112,23 +161,41 @@ module.exports = {
           });
 
           await newCoupon.save();
-          //return await Coupon.find({});
-          return { message: "Coupon saved successfully", status: 200 };
+          return {
+            message: Messages.AddSuccess.replace(":item", "Coupon"),
+            success: true,
+          };
         }
       } catch (error) {
-        //error = checkError(error);
-        return { message: `${errorRES.CREATE_ERROR} Coupon`, status: 400 };
+        return {
+          message: Messages.CREATE_ERROR.replace(":item", "Coupon"),
+          success: false,
+        };
       }
     },
     updateCoupon: async (root, args, { id }) => {
+      if (!id) {
+        return {
+          message: Messages.TOKEN_REQ.replace(":item", "Coupon"),
+          success: false,
+        };
+      }
       checkToken(id);
+      if (!args.id) {
+        return {
+          message: Messages.ID_ERROR.replace(":item", "Coupon"),
+          success: false,
+        };
+      }
       try {
         const coupon = await Coupon.findById({ _id: args.id });
         if (coupon) {
-          // Check Validation
           const errors = validate("updateCoupon", args);
           if (!isEmpty(errors)) {
-            throw putError(errors);
+            return {
+              message: errors,
+              success: false,
+            };
           }
 
           coupon.code = args.code;
@@ -146,30 +213,54 @@ module.exports = {
           coupon.updated = Date.now();
 
           await coupon.save();
-         // return await Coupon.find({});
-         return { message: "Coupon updated successfully", status: 200 };
-        } else {
-          //throw putError("Coupon not exist");
-          return { message: "Coupon not exist", status: 400 };
-        }
+          return {
+            message: Messages.UpdateSuccess.replace(":item", "Coupon"),
+            success: true,
+          };
+        } return {
+          message: Messages.NOT_EXIST.replace(":item", "Coupon"),
+          success: false,
+        };
       } catch (error) {
-        // error = checkError(error);
-        return { message: `${errorRES.UPDATE_ERROR} Coupon`, status: 400 };
+       error = checkError(error);
+       return {
+        message: Messages.UPDATE_ERROR.replace(":item", "Coupon"),
+        success: false,
+      };
       }
     },
     deleteCoupon: async (root, args, { id }) => {
+      if (!id) {
+        return {
+          message: Messages.TOKEN_REQ.replace(":item", "Coupon"),
+          success: false,
+        };
+      }
       checkToken(id);
+      if (!args.id) {
+        return {
+          message: Messages.ID_ERROR.replace(":item", "Coupon"),
+          success: false,
+        };
+      }
       try {
         const coupon = await Coupon.findByIdAndRemove(args.id);
         if (coupon) {
-          // const coupons = await Coupon.find({});
-          // return coupons || [];
-          return { message: "Coupon deleted successfully", status: 200 };
+          return {
+            message: Messages.DELETE.replace(":item", "Coupon"),
+            success: true,
+          };
         }
-        throw putError("customer not exist");
+        return {
+          message: Messages.NOT_EXIST.replace(":item", "Coupon"),
+          success: false,
+        };
       } catch (error) {
         error = checkError(error);
-        return { message: `${errorRES.DELETE_ERROR} Coupon`, status: 404 };
+        return {
+          message: Messages.DELETE_ERROR.replace(":item", "Coupon"),
+          success: false,
+        };
       }
     }
   }
